mlops-backend/services/dl_service/requirements.txt
requests>=2.31.0
numpy==1.25.2
pandas==1.4.2
Pillow==10.0.0
# tensorflow==2.10.0
tensorflow==2.10.0
opencv-python-headless==4.8.0.74
fastapi==0.101.1
pydantic>=1.2.0,<2.0.0
uvicorn[standard]==0.23.2
gunicorn==20.1.0
python-multipart==0.0.6
SQLAlchemy==1.4.49
psycopg2-binary==2.9.7
scikit-learn==0.24.2
torch==2.0.1
torchvision==0.15.2